Refactor ledger creation
