+++ src/core/providers/ProviderFactory.ts
@@ -2,15 +2,11 @@ import { CrossWindowProvider } from 'lib/sdkWebWalletCrossWindowProvider';
 import { ExtensionProvider } from '@multiversx/sdk-extension-provider';
 import {
   IProvider,
-  IProviderConfig,
   IProviderFactory,
   ProviderTypeEnum
 } from './types/providerFactory.types';
 import { isBrowserWithPopupConfirmation } from '../../constants';
-import { fetchAccount } from 'utils';
-import { setLedgerLogin } from 'store/actions/loginInfo/loginInfoActions';
-import { setLedgerAccount } from 'store/actions/account/accountActions';
-import { getLedgerProvider } from './helpers/getLedgerProvider';
+import { createLedgerProvider } from './helpers/ledger/createLedgerProvider';
 
 export class ProviderFactory {
   public async create({
@@ -49,105 +45,13 @@ export class ProviderFactory {
       }
 
       case ProviderTypeEnum.ledger: {
-        const data = await getLedgerProvider();
+        const ledgerProvider = await createLedgerProvider();
 
-        if (!data) {
+        if (!ledgerProvider) {
           return;
         }
 
-        const { ledgerProvider: provider, ledgerConfig } = data;
-
-        createdProvider = provider as unknown as IProvider;
-
-        const hwProviderLogin = provider.login;
-
-        createdProvider.getType = () => {
-          return ProviderTypeEnum.ledger;
-        };
-
-        createdProvider.login = async (options?: {
-          callbackUrl?: string | undefined;
-          token?: string | undefined;
-        }): Promise<{
-          address: string;
-          signature: string;
-        }> => {
-          const isConnected = provider.isConnected();
-
-          if (!isConnected) {
-            throw new Error('Ledger device is not connected');
-          }
-
-          // TODO: perform additional UI logic here
-          // maybe extract to file
-          const startIndex = 0;
-          const addressesPerPage = 10;
-
-          const accounts = await provider.getAccounts(
-            startIndex,
-            addressesPerPage
-          );
-
-          const accountsWithBalance: {
-            address: string;
-            balance: string;
-            index: number;
-          }[] = [];
-
-          const balancePromises = accounts.map((address) =>
-            fetchAccount(address)
-          );
-
-          const balances = await Promise.all(balancePromises);
-
-          balances.forEach((account, index) => {
-            if (!account) {
-              return;
-            }
-            accountsWithBalance.push({
-              address: account.address,
-              balance: account.balance,
-              index
-            });
-          });
-
-          // Suppose user selects the first account
-          const selectedIndex = 0;
-
-          setLedgerLogin({
-            index: selectedIndex,
-            loginType: ProviderTypeEnum.ledger
-          });
-
-          const { version, dataEnabled } = ledgerConfig;
-
-          setLedgerAccount({
-            address: accountsWithBalance[selectedIndex].address,
-            index: selectedIndex,
-            version,
-            hasContractDataEnabled: dataEnabled
-          });
-
-          if (options?.token) {
-            const loginInfo = await provider.tokenLogin({
-              token: Buffer.from(`${options?.token}{}`),
-              addressIndex: accountsWithBalance[selectedIndex].index
-            });
-
-            return {
-              address: loginInfo.address,
-              signature: loginInfo.signature.toString('hex')
-            };
-          } else {
-            const { address } = await hwProviderLogin({
-              addressIndex: accountsWithBalance[selectedIndex].index
-            });
-            return {
-              address,
-              signature: ''
-            };
-          }
-        };
+        createdProvider = ledgerProvider;
 
         break;
       }
