+++ src/core/providers/index.ts
@@ -1,2 +1,2 @@
 export { getAccountProvider } from './accountProvider';
-export * from './helpers/getLedgerConfiguration';
+export * from './helpers/ledger/getLedgerConfiguration';
